Update README.md
A company keeps its employee records in a file. Each line in the file includes a single employee record (id, first name, last name, & salary). Write a C++ program that prints a report of employee data. You are asked to process the data in the file by storing the records in a vector of structures. Your program must print the following report.

    All employee records sorted by last name.

    Find and print the total payroll.